+++ nbconvert_a11y/pytest_w3c.py
@@ -1,6 +1,7 @@
 # requires node
 # requires jvm
 import collections
+import dataclasses
 import functools
 import itertools
 import operator
@@ -22,7 +23,37 @@ import exceptiongroup
 import pytest
 import requests
 
-from nbconvert_a11y.axe.pytest_axe import Collector, Results, Violation
+from .exceptions import Violation
+from .axe.types import Base
+
+
+@dataclasses.dataclass
+class Collector(Base):
+    """the base collector class for accessibility and bulk testing batteries."""
+
+    url: str = None
+    results: Any = None
+
+    def configure(self):
+        return self
+
+    def exception(self):
+        return self.results.exception()
+
+    def raises(self):
+        exc = self.exception()
+        if exc:
+            raise exc
+
+
+class Results(Base):
+    data: Any
+
+    def raises(self):
+        exc = self.exception()
+        if exc:
+            raise exc
+
 
 HERE = Path(__file__).parent
 
@@ -149,8 +180,6 @@ def validate_html_file(a_vnu_server_url: str) -> TVnuValidator:
         return res.json()
 
     return post
-
-
 
 
 @pytest.fixture(scope="session")
